Continued test refactoring
