+++ dataland-frontend/tests/e2e/specs/backend-health/index.ts
@@ -2,4 +2,4 @@
  * As a developer, I want to ensure that these tests run against the latest version of the backend
  * and that the backend is up for the tests
  */
-import "./verifyDeployment";
+import "./VerifyDeployment";
