+++ dataland-frontend/tests/e2e/specs/company-search/SearchResults.functions.ts
@@ -1,7 +1,7 @@
 export function verifyTaxonomySearchResultTable(): void {
-    cy.get("table.p-datatable-table").contains("th", "COMPANY");
-    cy.get("table.p-datatable-table").contains("th", "PERM ID");
-    cy.get("table.p-datatable-table").contains("th", "SECTOR");
-    cy.get("table.p-datatable-table").contains("th", "MARKET CAP");
-    cy.get("table.p-datatable-table").contains("th", "LOCATION");
+  cy.get("table.p-datatable-table").contains("th", "COMPANY");
+  cy.get("table.p-datatable-table").contains("th", "PERM ID");
+  cy.get("table.p-datatable-table").contains("th", "SECTOR");
+  cy.get("table.p-datatable-table").contains("th", "MARKET CAP");
+  cy.get("table.p-datatable-table").contains("th", "LOCATION");
 }
