+++ dataland-frontend/tests/e2e/specs/company-search/index.ts
@@ -2,3 +2,4 @@
  * As a user I want to be able to search for companies and the search result page should behave as I expect it to
  */
 import "./Paginator"
+import "./IndexTabMenu"
