+++ dataland-frontend/tests/e2e/specs/prepopulation/Prepopulation.ts
@@ -1,4 +1,5 @@
-import { doThingsInChunks, retrieveDataIdsList, retrieveCompanyIdsList } from "../../utils/ApiUtils";
+import { retrieveDataIdsList, retrieveCompanyIdsList } from "../../utils/ApiUtils";
+import { doThingsInChunks } from "../../utils/Cypress";
 import { CompanyInformation, EuTaxonomyData } from "../../../../build/clients/backend/api";
 const chunkSize = 40;
