+++ dataland-frontend/tests/e2e/specs/user-authentication/Login.ts
@@ -1,5 +1,5 @@
 import Chainable = Cypress.Chainable;
-import { login, logout } from "./Login.functions";
+import { login, logout } from "../../utils/Auth";
 
 describe("As a user I want to be able to login and I want the login page to behave as I expect", () => {
   it("Checks that login works", () => {
