+++ dataland-frontend/tests/e2e/specs/user-authentication/index.ts
@@ -1,6 +1,8 @@
 /**
  * As a user I want to be able to login and I want the login page to behave as I expect
  * As a user I want to be able to register for an account and be able to log in and out of that account
+ * As a user I expect to be redirected to the login page if I am unauthenticated
  */
 import "./Login";
 import "./Register";
+import "./VisitPagesUnauth";
